user32: Extract check for our 32-to-64-bit thunks to a function.
Check alignment before attempting to read the magic number.  If the function
pointer isn't aligned, then a) we know for sure that it's not one of our thunks;
and b) it's not necessarily safe to read the memory that precedes it even though
the function itself may be in readable memory.

Check for the hotpatchable prolog.  If the 32-bit implementation has been
patched, we need to thunk to that rather than calling the 64-bit implementation.
Also, if the function pointer is invalid and points to unreadable memory,
triggering an exception by reading its first bytes is a closer approximation to
what would happen if we called it than reading bytes preceding it.  That may
matter for copy protection or anti-cheat software.

Finally, check for our magic number.